refactor: Consolidate schema imports and update validation messages in localization files
