+++ messages/pl.json
@@ -153,7 +153,8 @@
       },
       "email": {
         "incorrect": "Wpisz poprawny adres e-mail.",
-        "required": "E-mail jest wymagany."
+        "required": "E-mail jest wymagany.",
+        "max": "E-mail nie może mieć więcej niż {max} znaków."
       },
       "password": {
         "incorrect": "Wpisz poprawne hasło.",
