+++ src/app/api/auth/signUp/route.ts
@@ -1,7 +1,7 @@
 import { db } from '@/db/db';
 import { usersTable } from '@/db/schema';
-import type { SignUpValues } from '@/schemas';
-import { signUpSchema } from '@/schemas';
+import type { SignUpValues } from '@/schema';
+import { signUpSchema } from '@/schema';
 import { ApiError, handleApiError } from '@/utils/api';
 import { hashSync } from 'bcryptjs';
 import { eq, ilike, or } from 'drizzle-orm';
