+++ src/auth.ts
@@ -5,7 +5,7 @@ import Credentials from 'next-auth/providers/credentials';
 
 import { db } from './db/db';
 import { usersSelect, usersTable } from './db/schema';
-import { signInSchema } from './schemas';
+import { signInSchema } from './schema';
 import { ApiError } from './utils/api';
 
 export const { handlers, signIn, signOut, auth } = NextAuth({
