+++ src/components/organisms/AuthModal/AuthModalSignInForm/AuthModalSignInForm.tsx
@@ -3,8 +3,8 @@ import React from 'react';
 import { Box, Typography } from '@/components/atoms';
 import { useSignInMutation } from '@/hooks/api/auth/useSignInMutation';
 import { useAppForm } from '@/hooks/useFormHook';
-import type { SignInValues } from '@/schemas';
-import { signInSchema } from '@/schemas';
+import type { SignInValues } from '@/schema';
+import { signInSchema } from '@/schema';
 import { useAuthStore } from '@/stores/auth';
 import { useTranslations } from 'next-intl';
 import { useShallow } from 'zustand/shallow';
