+++ src/components/organisms/AuthModal/AuthModalSignUpForm/AuthModalSignUpForm.tsx
@@ -4,8 +4,8 @@ import { Box, Typography } from '@/components/atoms';
 import { useSignInMutation } from '@/hooks/api/auth/useSignInMutation';
 import { useSignUpMutation } from '@/hooks/api/auth/useSignUpMutation';
 import { useAppForm } from '@/hooks/useFormHook';
-import type { SignUpValues } from '@/schemas';
-import { signUpSchema } from '@/schemas';
+import type { SignUpValues } from '@/schema';
+import { signUpSchema } from '@/schema';
 import { useAuthStore } from '@/stores/auth';
 import { useTranslations } from 'next-intl';
 import { useShallow } from 'zustand/shallow';
