+++ src/schema/auth.ts
@@ -3,7 +3,7 @@ import { z } from 'zod';
 export const PASSWORD_MIN_LENGTH = 6;
 export const PASSWORD_MAX_LENGTH = 32;
 
-export const PROFILE_SCREEN_NAME_MIN_LENGTH = 3;
+export const PROFILE_SCREEN_NAME_MIN_LENGTH = 4;
 export const PROFILE_SCREEN_NAME_MAX_LENGTH = 15;
 export const EMAIL_MAX_LENGTH = 100;
 export const PROFILE_NAME_MIN_LENGTH = 4;
@@ -12,7 +12,7 @@ export const PROFILE_NAME_MAX_LENGTH = 50;
 export type SignInValues = z.infer<ReturnType<typeof signInSchema>>;
 export type SignUpValues = z.infer<ReturnType<typeof signUpSchema>>;
 
-export const signInSchema = (t?: Translation) =>
+export const signInSchema = (t: Translation | undefined = undefined) =>
   z.object({
     emailOrScreenName: z
       .string({ required_error: t && t('errors.validation.emailOrScreenName.required') })
@@ -29,7 +29,7 @@ export const signInSchema = (t?: Translation) =>
       ),
   });
 
-export const signUpSchema = (t?: Translation) =>
+export const signUpSchema = (t: Translation | undefined = undefined) =>
   z
     .object({
       screenName: z
@@ -44,12 +44,14 @@ export const signUpSchema = (t?: Translation) =>
         )
         .trim()
         .refine(
-          (value) => /^(\w+\S)*\w+$/g.test(value),
+          (value) => /^[a-zA-Z0-9][a-zA-Z0-9_]*$/g.test(value),
           t && t('errors.validation.screenName.invalid')
         ),
       email: z
         .string({ required_error: t && t('errors.validation.email.required') })
+        .toLowerCase()
         .email(t && t('errors.validation.email.incorrect'))
+        .max(EMAIL_MAX_LENGTH, t && t('errors.validation.email.max', { max: EMAIL_MAX_LENGTH }))
         .trim(),
       name: z
         .string({ required_error: t && t('errors.validation.name.required') })
@@ -63,7 +65,7 @@ export const signUpSchema = (t?: Translation) =>
         )
         .trim()
         .refine(
-          (value) => /^([a-zA-Z0-9]+\s)*[a-zA-Z0-9]+$/g.test(value),
+          (value) => /^[\p{L}0-9]+(?:[\s-][\p{L}0-9]+)*$/u.test(value),
           t && t('errors.validation.name.invalid')
         ),
       password: z
